requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
mlxtend>=0.21.0
xgboost>=1.7.0
shap>=0.41.0
pyyaml>=6.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
streamlit>=1.25.0
plotly>=5.15.0
openpyxl>=3.1.0
scipy>=1.10.0
